django/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.1"
gunicorn = "^20.0.4"
psycopg2-binary = "^2.8.6"
djangorestframework = "^3.11.1"
markdown = "^3.2.2"
django-cors-headers = "^3.5.0"
sentry-sdk = "^0.17.7"
djangorestframework-camel-case = "^1.2.0"
django-filter = "^2.3.0"
celery = "^5.0.1"
Pillow = "^9.0.1"
python-telegram-bot = "20.0a0"

[tool.poetry.dev-dependencies]
ipython = "^7.18.1"
ipdb = "^0.13.3"
pytest = "^6.0.2"
pytest-django = "^3.10.0"
pytest-cov = "^2.10.1"
pep8 = "^1.7.1"
autopep8 = "^1.5.4"
flake8 = "^3.8.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
